refactor(Modal): change component file name to kabab-case
