+++ src/components/modal/index.ts
@@ -1,3 +1,3 @@
-import Modal from './Modal';
+import Modal from './modal';
 
 export { Modal };
